third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-paint/timeline-paint-with-style-recalc-invalidations-expected.txt

Tests the Timeline API instrumentation of paint events with style recalc invalidations.


Running: testLocalFrame
first paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: test://evaluations/0/timeline-paint-with-style-recalc-invalidations.js:17}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "BODY"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: test://evaluations/0/timeline-paint-with-style-recalc-invalidations.js:18}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testSubframe
second paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: test://evaluations/0/timeline-paint-with-style-recalc-invalidations.js:24}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
